ml_model/requirements.txt
sklearn
snowflake-connector-python
python-dotenv
pandas
sklearn2pmml
